+++ assignment-2/src/main/java/com/lucentblock/assignment2/security/authentication/AuthenticationController.java
@@ -1,12 +1,10 @@
 package com.lucentblock.assignment2.security.authentication;
 
 import com.lucentblock.assignment2.security.model.*;
+import com.lucentblock.assignment2.service.SignupCodeService;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -32,16 +30,6 @@ public class AuthenticationController {
         String authHeader = request.getHeader("Authorization");
         String accessToken = authHeader.substring(7);
         return ResponseEntity.ok(authService.refresh(accessToken, refreshToken));
-    }
-
-    @PostMapping("/email-verification")
-    public ResponseEntity generateSignupCode(@Validated @RequestBody UserEmailDTO userEmailDTO) {
-        return authService.generateSignupCode(userEmailDTO.getUserEmail());
-    }
-
-    @PatchMapping("/email-verification")
-    public ResponseEntity verifySignupCode(@Validated @RequestBody VerifySignupCodeRequestDTO verifySignupCodeRequestDTO) {
-        return authService.verifySignupCode(verifySignupCodeRequestDTO);
     }
